messaging_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg==3.2.12
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0